+++ hyvideo/modules/attenion.py
@@ -28,7 +28,7 @@ def get_cu_seqlens(text_mask, img_len):
     max_len = text_mask.shape[1] + img_len
     device = text_mask.device
 
-    cu_seqlens = torch.zeros([2 * batch_size + 1], dtype=torch.int32, device=device)
+    cu_seqlens = torch.zeros([2 * batch_size + 1], dtype=torch.int16, device=device)
 
     for i in range(batch_size):
         s = text_len[i] + img_len
